% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_spdf_for_ggplot.r
\name{convert_spdf_for_ggplot2}
\alias{convert_spdf_for_ggplot2}
\title{convert_spdf_for_ggplot2}
\usage{
convert_spdf_for_ggplot2(df.sp, var_name = "var1.pred")
}
\arguments{
\item{df.sp}{a SpatialPointsDataFrame object (for more info on this class of object see the sp package)}

\item{var_name}{a string object specifying the column in the SpatialPointsDataFrame containing the variable to be
defined as z in the ggplot ready output data.frame}
}
\value{
df a data.frame object in long from for compatability with ggplot2; the var_name of interest is renamed z
}
\description{
A helper function that converts the SpatialPointDataFrame to a long-form output format
plottable in ggplot2. In the context of spcv it is useful for plotting the output of
gstat::idw(). The var_name argument defaults to "var1.pred" the column of interpolation predictions
produced by gstat::idw(). This variable is renamed to "z" in the output data.frame.
}
