% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.r
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(observations, predictions)
}
\arguments{
\item{observations}{vector of numeric values (NAs will be ignored)}

\item{predictions}{vector of numeric values (NAs will be ignored)}
}
\value{
the root mean squared error of two vectors
}
\description{
a simple root mean squared error calculator
}
