% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcv.r
\name{spcv}
\alias{spcv}
\title{spcv}
\usage{
spcv(df.sp, my_idp = 2, var_name = "conc", ...)
}
\arguments{
\item{df.sp}{a SpatialPointsDataFrame object See the sp package for how to create one.}

\item{my_idp}{a numeric object, specifying the inverse distance weighting bandwidth parameter}

\item{var_name}{a string object, specifying the variable name
for the data frame column containing numeric values to be interpolated.
For example, concentrations of a chemical found in spatial samples.}
}
\value{
list containing four objects: cv.input (SpatialPointsDataFrame),
cv.pred (vector of predictions at the hold out locations),
cv.error (vector of errors based on difference of hold out test points and interpolated prediction),
cv.rmse (numeric root mean squared error, which is vased on cv.error vector)
}
\description{
spcv
}
