% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_basic.r
\name{theme_basic}
\alias{theme_basic}
\title{theme_basic}
\usage{
theme_basic()
}
\value{
a ggplot2 theme object to produce a clean plot style in ggplot2 for maps
}
\description{
a ggplot2 theme with minimal contextual ink, for plotting spatial data with geom_points()
or geom_tile(). Ticks, axes labels, and legend title are omitted.
}
