% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_spdf_for_ggplot.r
\name{convert_spdf_for_ggplot2}
\alias{convert_spdf_for_ggplot2}
\title{convert_spdf_for_ggplot2}
\usage{
convert_spdf_for_ggplot2(df.sp, var_name = "var1.pred")
}
\arguments{
\item{df.sp}{}

\item{var_name}{}
}
\description{
convert_spdf_for_ggplot2
}
