% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.r
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(observations, predictions)
}
\arguments{
\item{observations}{a vector object of numeric values (NAs will be ignored)}

\item{predictions}{a vector object of numeric values (NAs will be ignored)}
}
\value{
the root mean squared error for the two vectors
}
\description{
calculate root mean squared error from two vectors
}
