% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcv.r
\name{spcv}
\alias{spcv}
\title{spcv}
\usage{
spcv(df.sp, my_idp = 2, var_name = "conc", ...)
}
\arguments{
\item{df.sp}{SpatialPointsDataFrame (see sp package)}

\item{my_idp}{numeric bandwidth parameter}

\item{var_name}{variable name of column containing numeric values (e.g., concentrations of a chemical found in samples)}
}
\value{
list containing cv.input (SpatialPointsDataFrame),
cv.pred (vector of predictions at hold out locations),
cv.error (vector of errors based on difference of test point - prediction),
cv.rmse (numeric root mean squared error based on cv.error)
}
\description{
spcv
}
