% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spcv.r
\name{spcv}
\alias{spcv}
\title{spcv}
\usage{
spcv(df.sp, my_idp = 2, var_name = "conc", ...)
}
\arguments{
\item{df.sp}{SpatialPointsDataFrame (see sp package)}

\item{my_idp}{numeric bandwidth parameter}

\item{var_name}{variable name of numeric}
}
\description{
spcv
}
